+++ Runtime/Core/Ease.cs
@@ -4,7 +4,7 @@ namespace FlowTween {
 
 /// <summary>
 /// Specifies a preset easing function.
-/// You can access the functions directly in <see cref="Easing"/>.
+/// You can access the functions directly in <see cref="Easings"/>.
 /// </summary>
 [Serializable]
 public struct Ease {
@@ -29,69 +29,69 @@ public struct Ease {
 
     /// <summary>
     /// Gets an easing function based on the current <see cref="Type"/> and <see cref="Direction"/>.
-    /// You can access them directly in <see cref="Easing"/>.
+    /// You can access them directly in <see cref="Easings"/>.
     /// </summary>
     public Func<float, float> GetFunction() {
         return Type switch {
-            EaseType.Linear => Easing.Linear,
+            EaseType.Linear => Easings.Linear,
             EaseType.Sine => Direction switch {
-                EaseDirection.In => Easing.SineIn,
-                EaseDirection.Out => Easing.SineOut,
-                EaseDirection.InOut => Easing.SineInOut,
+                EaseDirection.In => Easings.SineIn,
+                EaseDirection.Out => Easings.SineOut,
+                EaseDirection.InOut => Easings.SineInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Quad => Direction switch {
-                EaseDirection.In => Easing.QuadIn,
-                EaseDirection.Out => Easing.QuadOut,
-                EaseDirection.InOut => Easing.QuadInOut,
+                EaseDirection.In => Easings.QuadIn,
+                EaseDirection.Out => Easings.QuadOut,
+                EaseDirection.InOut => Easings.QuadInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Cubic => Direction switch {
-                EaseDirection.In => Easing.CubicIn,
-                EaseDirection.Out => Easing.CubicOut,
-                EaseDirection.InOut => Easing.CubicInOut,
+                EaseDirection.In => Easings.CubicIn,
+                EaseDirection.Out => Easings.CubicOut,
+                EaseDirection.InOut => Easings.CubicInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Quart => Direction switch {
-                EaseDirection.In => Easing.QuartIn,
-                EaseDirection.Out => Easing.QuartOut,
-                EaseDirection.InOut => Easing.QuartInOut,
+                EaseDirection.In => Easings.QuartIn,
+                EaseDirection.Out => Easings.QuartOut,
+                EaseDirection.InOut => Easings.QuartInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Quint => Direction switch {
-                EaseDirection.In => Easing.QuintIn,
-                EaseDirection.Out => Easing.QuintOut,
-                EaseDirection.InOut => Easing.QuintInOut,
+                EaseDirection.In => Easings.QuintIn,
+                EaseDirection.Out => Easings.QuintOut,
+                EaseDirection.InOut => Easings.QuintInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Expo => Direction switch {
-                EaseDirection.In => Easing.ExpoIn,
-                EaseDirection.Out => Easing.ExpoOut,
-                EaseDirection.InOut => Easing.ExpoInOut,
+                EaseDirection.In => Easings.ExpoIn,
+                EaseDirection.Out => Easings.ExpoOut,
+                EaseDirection.InOut => Easings.ExpoInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Circ => Direction switch {
-                EaseDirection.In => Easing.CircIn,
-                EaseDirection.Out => Easing.CircOut,
-                EaseDirection.InOut => Easing.CircInOut,
+                EaseDirection.In => Easings.CircIn,
+                EaseDirection.Out => Easings.CircOut,
+                EaseDirection.InOut => Easings.CircInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Back => Direction switch {
-                EaseDirection.In => Easing.BackIn,
-                EaseDirection.Out => Easing.BackOut,
-                EaseDirection.InOut => Easing.BackInOut,
+                EaseDirection.In => Easings.BackIn,
+                EaseDirection.Out => Easings.BackOut,
+                EaseDirection.InOut => Easings.BackInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Elastic => Direction switch {
-                EaseDirection.In => Easing.ElasticIn,
-                EaseDirection.Out => Easing.ElasticOut,
-                EaseDirection.InOut => Easing.ElasticInOut,
+                EaseDirection.In => Easings.ElasticIn,
+                EaseDirection.Out => Easings.ElasticOut,
+                EaseDirection.InOut => Easings.ElasticInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             EaseType.Bounce => Direction switch {
-                EaseDirection.In => Easing.BounceIn,
-                EaseDirection.Out => Easing.BounceOut,
-                EaseDirection.InOut => Easing.BounceInOut,
+                EaseDirection.In => Easings.BounceIn,
+                EaseDirection.Out => Easings.BounceOut,
+                EaseDirection.InOut => Easings.BounceInOut,
                 _ => throw new ArgumentOutOfRangeException()
             },
             _ => throw new ArgumentOutOfRangeException()