+++ Runtime/Factories/QuaternionTweenFactory.cs
@@ -5,7 +5,7 @@ namespace FlowTween {
 
 /// <summary>
 /// A factory for creating tweens that animate <see cref="Quaternion"/> properties.
-/// Also implements a composite interface for animating individual parts of
+/// Also implements a composite interface for animating individual components of
 /// the <see cref="Quaternion"/>, or Euler angles.
 /// </summary>
 /// <typeparam name="T">The type of the object that holds the property. Most commonly an <see cref="UnityEngine.Object"/>.</typeparam>
@@ -26,15 +26,15 @@ public class QuaternionTweenFactory<T> :
     /// than defining the getter and setter yourself. If performance is a concern, you should
     /// use <see cref="QuaternionTweenFactory{T}(System.Func{T,Quaternion},System.Action{T,Quaternion})"/> instead.
     /// </remarks>
-    public static QuaternionTweenFactory<T> Create(string propertyName) {
+    public static QuaternionTweenFactory<T> From(string propertyName) {
         return ReflectionTweenFactory.Create<QuaternionTweenFactory<T>>(propertyName);
     }
     
-    public void SetPart(ref Quaternion composite, Axis4 part, float value) => composite[(int)part] = value;
-    public float GetPart(Quaternion composite, Axis4 part) => composite[(int)part];
+    public void SetComponent(ref Quaternion composite, Axis4 component, float value) => composite[(int)component] = value;
+    public float GetComponent(Quaternion composite, Axis4 component) => composite[(int)component];
 
-    public void SetPart(ref Vector3 composite, Axis part, float value) => composite[(int)part] = value;
-    public float GetPart(Vector3 composite, Axis part) => composite[(int)part];
+    public void SetComponent(ref Vector3 composite, Axis component, float value) => composite[(int)component] = value;
+    public float GetComponent(Vector3 composite, Axis component) => composite[(int)component];
     
     public float Lerp(float from, float to, float t) => Mathf.LerpUnclamped(from, to, t);