+++ Runtime/Factories/Vector3TweenFactory.cs
@@ -5,7 +5,7 @@ namespace FlowTween {
 
 /// <summary>
 /// A factory for creating tweens that animate <see cref="Vector3"/> properties.
-/// Also implements a composite interface for animating individual parts of
+/// Also implements a composite interface for animating individual components of
 /// the <see cref="Vector3"/> by specifying one or more <see cref="Axis"/>.
 /// </summary>
 /// <typeparam name="T">The type of the object that holds the property. Most commonly an <see cref="UnityEngine.Object"/>.</typeparam>
@@ -25,18 +25,21 @@ public class Vector3TweenFactory<T> :
     /// than defining the getter and setter yourself. If performance is a concern, you should
     /// use <see cref="Vector3TweenFactory{T}(System.Func{T,Vector3},System.Action{T,Vector3})"/> instead.
     /// </remarks>
-    public static Vector3TweenFactory<T> Create(string propertyName) {
+    public static Vector3TweenFactory<T> From(string propertyName) {
         return ReflectionTweenFactory.Create<Vector3TweenFactory<T>>(propertyName);
     }
 
-    public void SetPart(ref Vector3 composite, Axis part, float value) => composite[(int)part] = value;
-    public float GetPart(Vector3 composite, Axis part) => composite[(int)part];
+    public void SetComponent(ref Vector3 composite, Axis component, float value) => composite[(int)component] = value;
+    public float GetComponent(Vector3 composite, Axis component) => composite[(int)component];
 
     public float Lerp(float from, float to, float t) => Mathf.LerpUnclamped(from, to, t);
     
     public ICompositeTweenFactory<Vector3, T, float, Axis> Composite => this;
 }
 
+/// <summary>
+/// Components of a Vector3.
+/// </summary>
 public enum Axis {
     X, Y, Z
 }